--- src/components/UserLogin/Login.vue
@@ -1,9 +1,0 @@
-<template>
-  <div>Login</div>
-</template>
-
-<script>
-export default {};
-</script>
-
-<style scoped lang="scss"></style>
