+++ src/views/Login/login.vue
@@ -1,0 +1,14 @@
+<template>
+  <UserLogin />
+</template>
+
+<script>
+import UserLogin from "@/components/UserLogin/UserLogin.vue";
+export default {
+  components: {
+    UserLogin,
+  },
+};
+</script>
+
+<style scoped lang="scss"></style>
